reorganize encryption errors
