+++ zk-token-sdk/src/encryption/mod.rs
@@ -13,5 +13,4 @@
 pub mod auth_encryption;
 pub mod discrete_log;
 pub mod elgamal;
-pub mod errors;
 pub mod pedersen;
